requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.104.1
h11==0.14.0
idna==3.6
joblib==1.3.2
newsapi-python==0.2.7
numpy==1.26.2
pydantic==2.5.2
pydantic_core==2.14.5
python-multipart==0.0.6
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
sklearn==0.0
sniffio==1.3.0
starlette==0.27.0
threadpoolctl==3.2.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
